refactor: consolidate duplicate normalization and company-ID logic
1. sbir_neo4j_loading: remove local normalize_company_name(), _LEGAL_SUFFIXES,
   and _ABBREVIATION_REPLACEMENTS; import normalize_name() from
   sbir_etl.utils.text_normalization with remove_suffixes=True and
   apply_abbreviations=True to preserve prior behavior.

2. sbir_ml/transition/analysis: extract duplicated _first_col() and
   _company_id_series() into _utils.py; merge column-candidate lists and
   null-value checks from both callers; both analytics.py and
   benchmark_evaluator.py now import from _utils.

3. text_normalization: fix broken relative import for ENHANCED_ABBREVIATIONS
   (module lives in sbir_etl.enrichers.matching, not sbir_etl.utils).
